+++ kotest-assertions/kotest-assertions-core/src/jvmTest/kotlin/com/sksamuel/kotest/matchers/doubles/DoubleToleranceTest.kt
@@ -2,8 +2,10 @@ package com.sksamuel.kotest.matchers.doubles
 
 import io.kotest.assertions.throwables.shouldThrowAny
 import io.kotest.core.spec.style.FunSpec
+import io.kotest.matchers.doubles.percent
 import io.kotest.matchers.doubles.plusOrMinus
 import io.kotest.matchers.shouldBe
+import io.kotest.matchers.shouldNotBe
 import io.kotest.property.Arb
 import io.kotest.property.arbitrary.numericDoubles
 import io.kotest.property.checkAll
@@ -21,6 +23,11 @@ class DoubleToleranceTest : FunSpec({
          Double.NEGATIVE_INFINITY shouldBe (Double.NEGATIVE_INFINITY plusOrMinus eps)
          Double.POSITIVE_INFINITY shouldBe (Double.POSITIVE_INFINITY plusOrMinus eps)
       }
+   }
+
+   test("Allow for percentage tolerance") {
+      1.5 shouldBe (1.0 plusOrMinus 50.percent)
+      1.5 shouldNotBe (2.0 plusOrMinus 10.percent)
    }
 })